+++ chakracore-cxx/lib/Jsrt/JsrtDebugUtils.cpp
@@ -45,7 +45,7 @@ void JsrtDebugUtils::AddLineColumnToObject(Js::DynamicObject* object, Js::Functi
 {
     if (functionBody != nullptr)
     {
-        ULONG line = 0;
+        uint32_t line = 0;
         int32_t col = 0;
         if (functionBody->GetLineCharOffset(byteCodeOffset, &line, &col, false))
         {
